Update actions and routes. Restrict actions for unauthorized users. Add data transformer for posts

+++ app/Http/Controllers/Api/PostController.php
@@ -8,9 +8,12 @@ use App\Models\Post;
 use App\Models\User;
 use Illuminate\Http\JsonResponse;
 use Illuminate\Http\Request;
+use Illuminate\Support\Facades\Gate;
 
 class PostController extends Controller
 {
+    const POSTS_PER_PAGE = 10;
+
     protected PostRepositoryInterface $postRepository;
 
     public function __construct(PostRepositoryInterface $postRepository)
@@ -21,9 +24,21 @@ class PostController extends Controller
     /**
      * Display a listing of the resource.
      */
-    public function index(): JsonResponse
+    public function index(Request $request): JsonResponse
     {
-        $posts = $this->postRepository->index();
+        $perPage = $request->query('per_page', self::POSTS_PER_PAGE); // Количество постов на страницу, по умолчанию 10
+        $posts = $this->postRepository->index($perPage);
+
+        // Трансформируем данные
+        $posts->getCollection()->transform(function ($post) {
+            return [
+                'id' => $post->id,
+                'user_id' => $post->user_id,
+                'title' => $post->title,
+                'url' => route('posts.show', ['id' => $post->id])
+            ];
+        });
+
         return response()->json($posts);
     }
 
@@ -36,6 +51,7 @@ class PostController extends Controller
     {
         $data = $request->all();
         $post = $this->postRepository->store($data);
+
         return response()->json($post, 201);
     }
 
@@ -47,6 +63,7 @@ class PostController extends Controller
     public function show(int $id): JsonResponse
     {
         $post = $this->postRepository->getById($id);
+
         return response()->json($post);
     }
 
@@ -58,8 +75,15 @@ class PostController extends Controller
      */
     public function update(Request $request, int $id): JsonResponse
     {
+        $post = $this->postRepository->getById($id);
+
+        if (Gate::denies('update', $post)) {
+            return response()->json(['error' => 'Unauthorized'], 403);
+        }
+
         $data = $request->all();
         $post = $this->postRepository->update($data, $id);
+
         return response()->json($post);
     }
 
@@ -70,18 +94,35 @@ class PostController extends Controller
      */
     public function destroy(int $id): JsonResponse
     {
+        $post = $this->postRepository->getById($id);
+
+        if (Gate::denies('delete', $post)) {
+            return response()->json(['error' => 'Unauthorized'], 403);
+        }
+
         $this->postRepository->delete($id);
+
         return response()->json(null, 204);
     }
 
     /**
      * @param User $user
-     * @param Request $request
      * @return JsonResponse
      */
-    public function userPosts(User $user, Request $request): JsonResponse
+    public function userPosts(User $user): JsonResponse
     {
-        $posts = $user->posts()->paginate(10); // Пагинация по 10 постов на страницу
+        $posts = $user->posts()->paginate(self::POSTS_PER_PAGE);
+
+        // Трансформируем данные
+        $posts->getCollection()->transform(function ($post) {
+            return [
+                'id' => $post->id,
+                'user_id' => $post->user_id,
+                'title' => $post->title,
+                'url' => route('posts.show', ['id' => $post->id])
+            ];
+        });
+
         return response()->json($posts);
     }
 }
